@@ -1,12 +1,20 @@
 ﻿using Microsoft.AspNetCore.Identity;
+using Microsoft.AspNetCore.Identity.UI.Services;
 
 namespace Boa.Identity;
 
-public abstract class User2FAServiceWithEmail<TUser> : User2FAServiceWithTokenProvider<TUser>
+public class User2FAServiceWithEmail<TUser> : User2FAServiceWithTokenProvider<TUser>
     where TUser : class
 {
-    public User2FAServiceWithEmail(IServiceProvider serviceProvider)
-        : base(serviceProvider, TokenOptions.DefaultEmailProvider) { }
+    private readonly IEmailSender _emailSender;
+
+    public User2FAServiceWithEmail(IServiceProvider serviceProvider, IEmailSender emailSender)
+        : base(serviceProvider, TokenOptions.DefaultEmailProvider)
+    {
+        _emailSender = emailSender;
+    }
+
+    public override bool NeedToSendToken => true;
 
     /// <summary>
     /// Asynchronously process the send token request after that the <paramref name="token"/> is generated.
@@ -27,20 +35,11 @@ public abstract class User2FAServiceWithEmail<TUser> : User2FAServiceWithTokenPr
             return false;
         }
 
-        await SendEmailAsync(
+        await _emailSender.SendEmailAsync(
             email,
             Localizer["Two Factor Authentication Token"],
             Localizer["Please, to authenticate use this code {0}", token]).ConfigureAwait(false);
 
         return true;
     }
-
-    /// <summary>
-    /// Sends an email to the specified <paramref name="email"/> address as an asynchronous operation.
-    /// </summary>
-    /// <param name="email">The email address to send email.</param>
-    /// <param name="subject">The subject of the eamil.</param>
-    /// <param name="htmlMessage">The body of the email in html format.</param>
-    /// <returns>A <see cref="Task"/> that completes when mail is sent.</returns>
-    protected abstract Task SendEmailAsync(string email, string subject, string htmlMessage);
 }
