@@ -1,3 +1,3 @@
-﻿@namespace Microsoft.AspNetCore.Identity.UI.V4.Pages.Internal
+﻿@namespace Boa.Identity.UI.V4.Pages.Internal
 
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
